+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2020-01-22
+
+### Changed
+- Success/Error ticks displayed in input
+- Show password displayed in inputs
+- Better password manager handling
+- Remove autofill browser styling
+
+
 ## [0.4.3] - 2020-01-18
 
 ### Changed